perf(kubernetes-react): load the pod exec terminal on demand
PodExecTerminal statically imported @xterm/xterm and its CSS, so every
consumer of plugin-kubernetes-react paid for xterm on first paint even
when no terminal was ever opened. Move the implementation behind
React.lazy + Suspense so xterm is fetched when a terminal actually
mounts, following the same pattern as OpenApiDefinitionWidget and
CodeSnippet. The public API is unchanged.
