+++ plugins/kubernetes-react/src/components/PodExecTerminal/PodExecTerminal.tsx
@@ -13,40 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import '@xterm/xterm/css/xterm.css';
 
-import { discoveryApiRef, useApi } from '@backstage/core-plugin-api';
-import { ClusterAttributes } from '@backstage/plugin-kubernetes-common';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
-import { useRef, useEffect, useMemo, useState } from 'react';
-import { Terminal } from '@xterm/xterm';
-import { FitAddon } from '@xterm/addon-fit';
+import { Progress } from '@backstage/core-components';
+import { lazy, Suspense } from 'react';
+import type { PodExecTerminalProps } from './PodExecTerminalContent';
 
-import { PodExecTerminalAttachAddon } from './PodExecTerminalAttachAddon';
+export type { PodExecTerminalProps } from './PodExecTerminalContent';
 
-/**
- * Props drilled down to the PodExecTerminal component
- *
- * @public
- */
-export interface PodExecTerminalProps {
-  cluster: ClusterAttributes;
-  containerName: string;
-  podName: string;
-  podNamespace: string;
-}
-
-const hasSocketProtocol = (url: string | URL) =>
-  /wss?:\/\//.test(url.toString());
-
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    podExecTerminal: {
-      width: '100%',
-      height: '100%',
-      '& .xterm-screen': { padding: theme.spacing(1) },
-    },
-  }),
+// @xterm/xterm and related CSS are large; only load them when a terminal mounts.
+const LazyPodExecTerminalContent = lazy(() =>
+  import('./PodExecTerminalContent').then(m => ({
+    default: m.PodExecTerminalContent,
+  })),
 );
 
 /**
@@ -54,88 +32,8 @@ const useStyles = makeStyles((theme: Theme) =>
  *
  * @public
  */
-export const PodExecTerminal = (props: PodExecTerminalProps) => {
-  const classes = useStyles();
-  const { containerName, podNamespace, podName } = props;
-
-  const [baseUrl, setBaseUrl] = useState(window.location.host);
-
-  const terminalRef = useRef(null);
-  const discoveryApi = useApi(discoveryApiRef);
-  const namespace = podNamespace ?? 'default';
-
-  useEffect(() => {
-    discoveryApi
-      .getBaseUrl('kubernetes')
-      .then(url => url ?? window.location.host)
-      .then(url => url.replace(/^http(s?):\/\//, 'ws$1://'))
-      .then(url => setBaseUrl(url));
-  }, [discoveryApi]);
-
-  const urlParams = useMemo(() => {
-    const params = new URLSearchParams({
-      container: containerName,
-      stdin: 'true',
-      stdout: 'true',
-      stderr: 'true',
-      tty: 'true',
-      command: '/bin/sh',
-    });
-    return params;
-  }, [containerName]);
-
-  const socketUrl = useMemo(() => {
-    if (!hasSocketProtocol(baseUrl)) {
-      return '';
-    }
-
-    return new URL(
-      `${baseUrl}/proxy/api/v1/namespaces/${namespace}/pods/${podName}/exec?${urlParams}`,
-    );
-  }, [baseUrl, namespace, podName, urlParams]);
-
-  useEffect(() => {
-    if (!hasSocketProtocol(socketUrl)) {
-      return () => {};
-    }
-
-    const terminal = new Terminal();
-    const fitAddon = new FitAddon();
-    terminal.loadAddon(fitAddon);
-
-    if (terminalRef.current) {
-      terminal.open(terminalRef.current);
-      fitAddon.fit();
-    }
-
-    terminal.writeln('Starting terminal, please wait...');
-
-    const socket = new WebSocket(socketUrl, ['channel.k8s.io']);
-
-    socket.onopen = () => {
-      terminal.clear();
-      const attachAddon = new PodExecTerminalAttachAddon(socket, {
-        bidirectional: true,
-      });
-
-      terminal.loadAddon(attachAddon);
-    };
-
-    socket.onclose = () => {
-      terminal.writeln('Socket connection closed');
-    };
-
-    return () => {
-      terminal?.clear();
-      socket?.close();
-    };
-  }, [baseUrl, socketUrl]);
-
-  return (
-    <div
-      data-testid="terminal"
-      ref={terminalRef}
-      className={classes.podExecTerminal}
-    />
-  );
-};
+export const PodExecTerminal = (props: PodExecTerminalProps) => (
+  <Suspense fallback={<Progress />}>
+    <LazyPodExecTerminalContent {...props} />
+  </Suspense>
+);
